requirements.txt
alabaster==1.0.0
asgiref==3.9.1
babel==2.17.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.4
djangorestframework==3.15.2
docutils==0.21.2
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.5
oauthlib==3.3.1
packaging==25.0
Pygments==2.19.2
python-decouple==3.8
requests==2.32.4
requests-oauthlib==2.0.0
roman-numerals-py==3.1.0
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
